+++ ZDZCode.Payments.Asaas/ValueObject/Ted.cs
@@ -1,5 +1,5 @@
 ﻿namespace ZDZCode.Payments.Asaas.ValueObject
 {
-    public record Ted(decimal FeeValue, bool ConsideredInMonthlyTransfersWithoutFee);
+    public record Ted(double FeeValue, bool ConsideredInMonthlyTransfersWithoutFee);
     
 }
